Shop main layout established, small changes in other files
