+++ Tavern/Data/Carts.cs
@@ -1,0 +1,7 @@
+namespace Tavern.Data {
+    public class Carts
+    {
+        public required string Cart_ID {get; set;}
+        public double? Price {get; set;}
+    }
+}
