--- Tavern/Data/Carts_TableClass.cs
@@ -1,5 +1,0 @@
-public class Carts
-{
-    public required string Cart_ID {get; set;}
-    public double? Price {get; set;}
-}
