+++ Tavern/Data/ContactList.cs
@@ -1,0 +1,8 @@
+namespace Tavern.Data {
+    public class ContactList
+    {
+        public required string CLI_ID {get; set;}
+        public string? Last_Name {get; set;}
+        public string? Email {get; set;}
+    }
+}
