--- Tavern/Data/Contact_List_TableClass.cs
@@ -1,6 +1,0 @@
-public class Contact_List
-{
-    public required string CLI_ID {get; set;}
-    public string? Last_Name {get; set;}
-    public string? Email {get; set;}
-}
