+++ Tavern/Data/Orders.cs
@@ -1,0 +1,9 @@
+namespace Tavern.Data {
+    public class Orders
+    {
+        public required string Order_ID {get; set;}
+        public string? Email {get; set;}
+        public string? Address {get; set;}
+        public bool? Delivered {get; set;}
+    }
+}
