--- Tavern/Data/Orders_TableClass.cs
@@ -1,7 +1,0 @@
-public class Orders
-{
-    public required string Order_ID {get; set;}
-    public string? Email {get; set;}
-    public string? Address {get; set;}
-    public bool? Delivered {get; set;}
-}
