--- Tavern/Data/Product_TableClass.cs
@@ -1,11 +1,0 @@
-public class Product
-{
-    public required string Product_ID {get; set;}
-    public string? Item_Type {get; set;}
-    public int? Price {get; set;}
-    public string? Item_Name {get; set;}
-    public string? Image {get; set;}
-    public string? Item_Desc {get; set;}
-    public int? Stock_num {get; set;}
-    public string? Color {get; set;}
-}
